+++ cmd/clicmd/startserver.go
@@ -3,7 +3,6 @@ package cmd
 import (
 	"fmt"
 	"log"
-	"os"
 	"path/filepath"
 	"runtime"
 
@@ -36,16 +35,14 @@ func serverStart(_ *cobra.Command, _ []string) {
 	numCPU := runtime.NumCPU()
 	runtime.GOMAXPROCS(numCPU)
 	configLog(stdoutLog)
-	//create temp folder for exported file
-	err := os.MkdirAll("./exported", 0730)
-	if err != nil {
-		panic(err)
-	}
 	//get configuration from ENV variable
 	kyberENV := common.RunningMode()
 	InitInterface()
 	config := GetConfigFromENV(kyberENV)
-	backupLog(config.Archive)
+	//backup other log daily
+	backupLog(config.Archive, "@daily", "core.+\\.log")
+	//backup core.log every 2 hour
+	backupLog(config.Archive, "@every 2h", "core\\.log")
 
 	var (
 		rData reserve.ReserveData
@@ -55,21 +52,21 @@ func serverStart(_ *cobra.Command, _ []string) {
 	)
 	//Create Data and Core, run if not in dry mode
 	if !noCore {
-		var iErr error
+		var err error
 		//create blockchain only if there is core
-		bc, iErr = CreateBlockchain(config, kyberENV)
-		if iErr != nil {
-			log.Panicf("Can not create blockchain: (%s)", iErr)
+		bc, err = CreateBlockchain(config, kyberENV)
+		if err != nil {
+			log.Panicf("Can not create blockchain: (%s)", err)
 		}
 		rData, rCore = CreateDataCore(config, kyberENV, bc)
 		if !dryrun {
 			if kyberENV != common.SimulationMode {
-				if iErr = rData.RunStorageController(); iErr != nil {
+				if err = rData.RunStorageController(); err != nil {
 					log.Panic(err)
 				}
 			}
-			if iErr = rData.Run(); iErr != nil {
-				log.Panic(iErr)
+			if err = rData.Run(); err != nil {
+				log.Panic(err)
 			}
 		}
 		//set static field supportExchange from common...