+++ cmd/clicmd/startserver_functions.go
@@ -44,15 +44,15 @@ var (
 	stagingOldNetwork = [1]string{"0xD2D21FdeF0D054D2864ce328cc56D1238d6b239e"}
 )
 
-func backupLog(arch archive.Archive) {
+func backupLog(arch archive.Archive, cronTimeExpression string, fileNameRegrexPattern string) {
 	c := cron.New()
-	err := c.AddFunc("@daily", func() {
+	err := c.AddFunc(cronTimeExpression, func() {
 		files, rErr := ioutil.ReadDir(logDir)
 		if rErr != nil {
 			log.Printf("BACKUPLOG ERROR: Can not view log folder - %s", rErr.Error())
 		}
 		for _, file := range files {
-			matched, err := regexp.MatchString("core.*\\.log", file.Name())
+			matched, err := regexp.MatchString(fileNameRegrexPattern, file.Name())
 			if (!file.IsDir()) && (matched) && (err == nil) {
 				err := arch.UploadFile(arch.GetLogBucketName(), remoteLogPath, filepath.Join(logDir, file.Name()))
 				if err != nil {
@@ -186,6 +186,7 @@ func CreateDataCore(config *configuration.Config, kyberENV string, bc *blockchai
 	//get fetcher based on config and ENV == simulation.
 	dataFetcher := fetcher.NewFetcher(
 		config.FetcherStorage,
+		config.StepFunctionDataStorage,
 		config.FetcherGlobalStorage,
 		config.World,
 		config.FetcherRunner,
@@ -202,6 +203,7 @@ func CreateDataCore(config *configuration.Config, kyberENV string, bc *blockchai
 	dataFetcher.SetBlockchain(bc)
 	rData := data.NewReserveData(
 		config.DataStorage,
+		config.StepFunctionDataStorage,
 		dataFetcher,
 		config.DataControllerRunner,
 		config.Archive,